First Major Commit, Built and installed opencv both with and without cuda/cudnn support.
I have bundled in the cpu only build under third-party. See LICENSES for license information.

In CMakePresets.json you have 4 ways to build the code. cpu-debug/cpu-relase/cuda-debug/cuda-release. All of which will collect the necessary dll files into the build directory along with the executable. This makes it to where the code can quickly be bundled up into an installer or zip file for distribution.

I do not have distribution rights for the CUDA project, so If you wish to run this project with CUDA support you will need to build your own copy of opencv using cuda. And point CMake to that install directory in the CMakeLists.txt file.

This will be the starting point for the project, future commits will either address bugs found in this code, or add features ontop of what exists now.