Split the program into several files. video_capture, video_display, apple_detection, and utils.
These each run their own threads to better separate everything out so that the program is less of an inline process and more like a small factory.

The main.cpp:main() function now handles startup. Startup calls video_capture, to prompt the user to select a device. Then this output gets sent directly to apple_detection for now. No other modes, others will be added back later.

Apple detection now has three threads, pre-process, forward, and post-process. Pre-process takes a frame from the video capture device, and reshapes it to fit the model. Then generates the blob to be passed to forward. Forward takes a blob, and runs detections on it, and passes the detections to post-process. Post process is where all the magic will happen in terms of maintaining gamestate.

Then the display thread will handle the display, screenshot, recording features. and automatically displays the fps and a diagnostic line that can be dynamically written.

The inputs registered to the display window get passed back to main in a callback function for future work with other modes and features.