IT FINALLY WORKS!!!!!
Moved away from the velocity checking, and started looking instead to see what apples are moving. And staging the swap when they leave the "origin point" once the apples re-enter the origin circle, apply the swap. This enables it to run alot faster, and way more precise.

Minor issue, I need to start tracking the frame of the game window. So that if the apple origins move I can update their positions as the frame moves.

As long as I can see 3 of the corners, I can "triangulate" the origin points from their starting position. And update it per game update.

This will take some time and will require dataset updates to start tracking the corners of the frame. Should only affect when using the webcam.